Changed default config; Added /join command; Changed OBItem lore; Fixed bug with Turrets; Added OBItem Mjolnir; Fixed function to give skull a texture

+++ src/main/java/me/ferdithedev/overblock/games/arena/Cuboid.java
@@ -55,7 +55,7 @@ public class Cuboid {
     }
 
     public Location getCenter() {
-        return new Location(this.world, (this.xMax - this.xMin) / 2 + this.xMin, (this.yMax - this.yMin) / 2 + this.yMin, (this.zMax - this.zMin) / 2 + this.zMin);
+        return new Location(this.world, (double) (this.xMax - this.xMin) / 2 + this.xMin, (double) (this.yMax - this.yMin) / 2 + this.yMin, (double) (this.zMax - this.zMin) / 2 + this.zMin);
     }
 
     public double getDistance() {
